requirements.txt
openai==1.40.0
python-dotenv==1.0.1
python-telegram-bot==20.7
